requirements.txt
addict==2.4.0
aiodns==4.0.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
altair==6.0.0
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.11.0
APScheduler==3.10.4
asgiref==3.11.1
async-generator==1.10
attrs==25.4.0
Authlib==1.6.6
babel==2.18.0
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==5.0.0
beartype==0.22.9
beautifulsoup4==4.12.3
bidict==0.23.1
black==26.3.1
blinker==1.9.0
blis==1.3.3
build==1.4.0
cachetools==6.2.6
catalogue==2.0.10
ccxt==4.5.40
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.2
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
coincurve==21.0.0
colorama==0.4.6
colorlog==6.10.1
confection==0.1.5
contourpy==1.3.3
cryptography==43.0.1
csvw==3.7.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cyclopts==4.5.1
cymem==2.0.13
dataclass-wizard==0.39.1
decorator==4.4.2
Deprecated==1.3.1
diskcache==5.6.3
distro==1.9.0
dlinfo==2.0.0
dnspython==2.8.0
docopt==0.6.2
docstring_parser==0.17.0
docutils==0.22.4
duckduckgo_search==6.3.7
durationpy==0.10
ecdsa==0.19.1
edge-tts==7.2.7
elevenlabs==1.9.0
email-validator==2.3.0
espeakng-loader==0.2.4
exceptiongroup==1.3.1
fakeredis==2.33.0
Farama-Notifications==0.0.4
fastapi==0.115.0
fastmcp==2.14.4
fastuuid==0.14.0
filelock==3.20.0
flake8==7.3.0
flatbuffers==25.12.19
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.147.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==1.1.2
grpcio==1.78.0
grpcio-status==1.62.3
gunicorn==25.1.0
gymnasium==1.2.2
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hf-xet==1.3.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx>=0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
importlib_metadata==8.4.0
importlib_resources==6.5.2
iniconfig==2.3.0
iqoptionapi==0.5
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.2
jose==1.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jsonwebtoken==1.0.0
keyring==25.7.0
kiwisolver==1.4.9
kokoro==0.9.4
kubernetes==35.0.0
language-tags==1.2.0
litellm==1.81.16
loguru==0.7.3
lupa==2.6
lxml==5.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mccabe==0.7.0
mcp==1.26.0
mdurl==0.1.2
misaki==0.9.4
mmh3==5.2.0
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.13.0
networkx==3.6
notebooklm-mcp-server==0.1.15
notion-client==2.2.1
num2words==0.5.14
numpy==2.4.3
numpy-quaternion==2024.0.13
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.24.2
openai==2.30.0
openapi-pydantic==0.5.1
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.11.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathable==0.4.4
pathspec==1.0.4
pathvalidate==3.3.1
pefile==2024.8.26
phonemizer-fork==3.3.2
pillow==12.1.1
platformdirs==4.5.1
playwright==1.58.0
plotly==6.5.0
pluggy==1.6.0
posthog==5.4.0
preshed==3.0.12
primp==1.1.3
proglog==0.1.12
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.25.8
psutil==7.2.2
psycopg2-binary==2.9.11
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyaes==1.6.1
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycares==5.0.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pydocket==0.16.6
pydub==0.25.1
pyee==13.0.1
pyflakes==3.4.0
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PyJWT==2.10.1
PyMsgBox==2.0.1
pyngrok==7.5.0
pyparsing==3.2.5
PyPDF2==3.0.1
pyperclip==1.11.0
PyPika==0.51.1
pypng==0.20220715.0
pyproject_hooks==1.2.0
pyqtgraph==0.14.0
pyquaternion==0.9.9
PyRect==0.2.0
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
PySocks==1.7.1
pytchat==0.5.5
pytesseract==0.3.13
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.2.2
python-engineio==4.12.3
python-jose==3.5.0
python-json-logger==4.0.0
python-multipart==0.0.12
python-socketio==5.15.0
python-telegram-bot==22.5
pytokens==0.4.1
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.3
qrcode==7.4.2
rdflib==7.6.0
redis==5.1.0
referencing==0.36.2
regex==2026.2.28
replicate==0.34.2
requests==2.33.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==14.3.1
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.16.3
segments==2.4.0
selenium==4.40.0
sentence-transformers==5.2.3
shellingham==1.5.4
shiboken6==6.10.1
simple-websocket==1.1.0
six==1.17.0
slack_sdk==3.33.0
smart_open==7.5.1
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.8.3
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.15.2
SQLAlchemy==2.0.45
sqlmodel==0.0.27
srsly==2.5.2
sse-starlette==3.2.0
stable_baselines3==2.7.0
starlette==0.38.6
streamlit==1.55.0
structlog==24.4.0
sympy==1.14.0
ta==0.11.0
tabulate==0.10.0
tavily-python==0.5.0
Telethon==1.42.0
tenacity==9.1.4
termcolor==3.3.0
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
tornado==6.5.5
tqdm==4.67.3
transformers==5.2.0
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
tweepy==4.14.0
typer==0.24.1
typer-slim==0.24.0
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.30.6
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.7.0
weasel==0.4.3
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.3.2
xgboost==3.1.2
yarl==1.22.0
zipp==3.23.0
zod==0.8.0
